+++ changelog.md
@@ -184,6 +184,8 @@ proc mydiv(a, b): int {.raises: [].} =
 - Added the `thiscall` calling convention as specified by Microsoft, mostly for hooking purpose
 - Deprecated `{.unroll.}` pragma, was ignored by the compiler anyways, was a nop.
 
+- Removed `asyncdispatch.newAsyncNativeSocket`, was deprecated since `0.18`.
+
 
 ## Compiler changes